backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9

# --- AI & NLP ECOSYSTEM ---
langchain>=0.1.13
langgraph>=0.0.30
langchain-openai>=0.1.1

# --- OCR & DOCUMENT PARSING (LIGHTWEIGHT) ---
# PyMuPDF extrai texto e bbox de PDFs sem dor de cabeça de ML
PyMuPDF==1.24.0
unstructured>=0.12.6
python-docx>=1.1.0
pypdf>=4.0.0
pdf2image>=1.17.0
pytesseract>=0.3.10
python-magic>=0.4.27

# --- EXPORT & PIPELINE B2B ---
reportlab>=4.1.0
pandas>=2.2.0
openpyxl>=3.1.2

# --- AUTH ---
firebase-admin>=6.4.0

# --- E-SIGNATURE (ZAPSIGN) ---
httpx>=0.27.0
